ci: Continuous Integration for Windows OBS Deps
This adds the necessary scripts to completely build the Windows dependencies from scratch on Continuous Integration (and possible home systems) with ease. The scripts can generate 32-bit and 64-bit dependencies and have various options.

In addition to that, this also adds the necessary configuration for Github Actions which will deal with the compile, package and artifact step for us. Github Actions are several times faster than Travis, Circle-CI and AppVeyor, and as such are ideal for this purpose.

The supported Modules are:
- zlib: 1.2.11
- mbedtls: 2.19.1 with MSVC fixes
- ogg: 1.3.4
- vorbis: 9eadeccdc4247127d91ac70555074239f5ce3529 (1.3.6 has a vulnerability)
- opus: 1.3.1
- vpx: 1.8.1
- libpng: 1.6.37
- x264: 3759fcb7b48037a5169715ab89f80a0ab4801cdf (no version tags)
- nv-codec-headers: 9.1.23.0
- srt: 1.4.0
- ffmpeg: 4.2.1
- curl: 7.66.0
- luajit: 2.0.5 with CMake support
- python: 3.7.5
- swig: 3.0.12