Add CEF patch file
This adds the CEF patch file used for OSR support on CEF version 4430 that OBS
requires for hardware acceleration to work properly. It is being committed here
for tracking and transparency, and if we need any future patches they can be
maintained here instead of pulled from a gist from a link in a PR that is
closed.