Fix github_fetch argument count check
To add support for git sparse-checkout, I had tried to change
github_fetch to allow it to accept three OR four arguments. Instead of
causing it to error out if less than three arguments were supplied, I'd
caused it to error out if three or less arguments weere supplied. This
is because I'd used the -le operator (less than or equal to) instead of
the -lt operator (less than). Use the correct -le operator here.