windows: Build AJA debug library
The OBS AJA plugin (aja) currently requires both the Release and Debug
versions of the ajantv2 library to be able to build. This enables CI to
build both Release and Debug versions of the library. Local builds with
the build-deps-windows-native.ps1 script will still only produce a
single build in the RelWithDebInfo configuration.