CI: Improve Qt caching
Use the hash of the Qt patches for the cache key for the Qt build jobs.
Use restore-keys to do a fuzzy-match to restore the cache if the exact
cache key is not found. This should allow us to restore the Qt cache and
do a rebuild if the patches used for Qt change. This means we don't have
to remember to change the QT_REVISION variable if we add new patches or
change the existing ones.