Fix broken encrypted SRT connections
The patches that were being applied to libsrt in cross-compiled builds
were causing encrypted SRT connections (such as to Cloudflare) to fail
to establish a successful connection during the handshake. These patches
were only meant to be used in the native-compiled static build of
mbedTLS. Removing this change from the cross-compiled build resolves the
issue.