Add zsh-based build framework for FFmpeg
With zsh being the default interactive shell on macOS, this future-
proofs the build script for all platforms (as zsh is easily installed
on Linux systems as well).

Using zsh allows using more powerful globbing and parameter substitution
patterns as well as modular functions available for autoloading in build
scripts.

Minimum required zsh version is 5.2 (until proven otherwise).

Functionality includes (but is not limited to):
* Applying patches (local or from URL) with checksum verification
* Downloading with automatic checksum verification
* Git checkouts with support for branches, GitHub-based pull-requests
and sparse checkouts
* Fixing rpaths of shared libraries (macOS)
* Creating import libraries for shared libraries (Windows)
* Cross-compilation (Windows) and native-compilation on Linux