libpng: Fix shared library being shipped with macOS dependencies
libpng is just needed for FFmpeg and freetype as a static library. As
the non-FFmpeg-deps are built as shared libraries, libpng got built as
a shared library as well and thus shipped with the package.

This commit forces libpng to be built as a static library, so freetype
can pick it up, but it will be removed before packaging.