utils.pwsh: Add Test-CommandExists
Get-Command will fail hard if the specified command is not found. This
means that attempts to automatically install commands/software if they
are not present on the system will currently cause the entire script to
stop executing immediately. Test-CommandExists allows us to gracefully
detect the presence of a command and take further action based on
whether or not it is found.