Qt5: Skip architecture thinning for CI runs
When cross-compiling Qt5 on CI, binaries should not be thinned, as the
artifacts will be used by CI runners to compile OBS in a later step.

For that reason, binaries need to stay universal (including QtCore,
which is linked to by some binaries) so they can run on x86_64 runners.