deps.windows: Add Jansson to Windows deps
Add Jansson 2.14 to the Windows deps. Mimic our existing in-tree build
options as much as possible, except don't override JANSSON_STATIC_CRT.
The default value is OFF, which in turn means /MT and /MTd are NOT set,
which is seemingly what we want. Our in-tree build overrides
JANSSON_STATIC_CRT to ON, but then also forces /MT and /MTd to NOT be
set. JANSSON_STATIC_CRT only controls whether or not those flags are
set, so just leave it OFF.