deps.windows: Explicitly disable ZLIB with curl
Previously, curl would attempt to find ZLIB, and could continue even if
ZLIB was not fully found. This could result in a scenario where ZLIB
includes were found, but ZLIB libraries were not found, and the curl
build would continue and succeed anyway. Examining previous builds, it
seems that this was the case for our dependencies.

A recent build failure on CI (windows-2022, version 20221120.1) had
CMake find the x64 ZLIB libraries from CI's Strawberry Perl install
location, which caused a build failure for the x86 dependencies due to
mismatched architectures. The version of Perl did not change between the
image versions, so I suspect some other change, possibly in CMake, that
caused it to suddenly find the libraries.

Since ZLIB was not actually being used, let's simply disable it
explicitly.

References:
https://github.com/curl/curl/commit/f21cc62832a9d5c521d10da9177b075bfed43f2f
https://github.com/actions/runner-images/blob/win22/20221027.1/images/win/Windows2022-Readme.md
https://github.com/actions/runner-images/blob/win22/20221120.1/images/win/Windows2022-Readme.md