deps.ffmpeg: Fix AOM target CPU arch
AOM allows you to set the target CPU arch (AOM_TARGET_CPU) at configure
time. If not specified, their CMake will attempt to autodetect this
value based on CMAKE_SYSTEM_PROCESSOR. The value of AOM_TARGET_CPU is
then used to enable various optimizations.

We used to not set a value for AOM_TARGET_CPU.[1] We started setting a
value for it with the unified FFmpeg build. However, we set the value of
"arch" based on the build target string (i.e., windows-x86 or
windows-x64), which results in an "arch" of "x86" or "x64". AOM's CMake
does expect the value "x64" as a possible value of AOM_TARGET_CPU, as it
instead expects "x86_64".[2] However, it will not produce a configure
error if given "x64" for the value of AOM_TARGET_CPU. It will accept
this value, but then subsequent checks will rely on the value being
"x86" or "x86_64" for things like using YASM or NASM.

AOM's CMake will correctly set this from the value of
CMAKE_SYSTEM_PROCESSOR, which we already set for Windows builds, so
let's just remove AOM_TARGET_CPU as a global override option, and
instead also specify CMAKE_TOOLCHAIN_FILE for macOS builds.


References:
[1]: https://github.com/obsproject/obs-deps/blob/6366c22b731c31112449d1d323daafaf8e306e3c/CI/windows/build_libaom.sh#L28-L43
[2]: https://aomedia.googlesource.com/aom.git/+/refs/tags/v3.6.0/build/cmake/aom_configure.cmake#57
 *   https://cmake.org/cmake/help/v3.26/variable/CMAKE_SYSTEM_PROCESSOR.html
 *   https://aomedia.googlesource.com/aom.git/+/refs/tags/v3.4.0/build/cmake/aom_configure.cmake#53