deps.windows: Re-enable mbedTLS
This enables mbedTLS to be compiled natively on MSVC.
The lib is built as a static dependency because mbedTLS dlls can not
be built natively with MSVC due to long-standing bugs.

Note: for future upgrades, 3.3.0 has a bug with designated initializers
which prevents the building of libdatachannel on MSVC. But 3.4.0 does
not.

The script was ripped from [1] authored by PatTheMav.

[1] https://github.com/obsproject/obs-deps/pull/186

Signed-off-by: pkv <pkv@obsproject.com>
Co-authored-by: PatTheMav <PatTheMav@users.noreply.github.com>
Signed-off-by: pkv <pkv@obsproject.com>