CI: Enable qrcodegencpp Debug build
Building obs-studio/obs-websocket against a Release/RelWithDebInfo
qrcodegencpp library fails because of mismatched values for
_ITERATOR_DEBUG_LEVEL and RuntimeLibrary. To fix this, we can build
qrcodegencpp in Debug and ship both Debug and RelWithDebInfo copies.
However, the qrcodegen-cmake project doesn't automatically set a debug
build suffix/postfix for qrcodegencpp. To fix that, we can manually set
this via CMAKE_DEBUG_POSTFIX so that the Release/RelWithDebInfo and
Debug versions don't clobber each other when putting them in the same
directory.

Enable building in Debug on CI so that both versions are built, similar
to other libraries that have this requirement.