deps.qt: Restore native dialogs on macOS
Revert qtbase 1b71e2d894c2be7052518cdcb96020c9950e2dc7:
QMessageBox: don't use a native dialog if a style sheet is active

The commit in question causes us to lose native dialogs on macOS. We can
restore them by reverting the above commit. This is hopefully a
temporary measure until either this is addressed in Qt itself or we get
guidance on the correct way to deal with this.