deps.ffmpeg: Remove MbedTLS CMake files from macOS FFmpeg package
Once the project switches MbedTLS 3.6+, there needs to be a way to
distinguish between older and newer versions on the obs-studio side.

The path of least resistance is to use CMake package-based discovery
for versions 3.6+ and find module-based discovery for older versions,
which requires the older (current as of time of the commit) versions
to not provide a CMake package as part of obs-deps.