deps.ffmpeg: Update SVT-AV1 to 3.0.0
SVT-AV1 3.0.0 no longer vendors cpuinfo. They now use CMake's
FetchContent by default, unless we specify at configure time to use an
external copy. This means that we cannot Patch cpuinfo before Configure.

If we determine that the patch is no longer required, then we can simply
remove it.

If we determine that the patch is still required, then we have to
either:
 A. upstream our patch, then get SVT-AV1 to update the commit they use
    for FetchContent
 B. figure out if we can use a SYSTEM/EXTERNAL cpuinfo
 C. disable cpuinfo (USE_CPUINFO=OFF)