deps.qt: Enable creation of dSYM files for all configurations
On macOS debug information is kept in the object files generated by the
compiler and extracted into dSYM files using the reference information
kept in the binaries.

Any binary built in debug information will not contain any debug
information, instead the toolchain will find the information in the
object files locally. Thus binaries created in debug configuration need
to be shipped with their associated dSYM files if they are to be used
on a different machine.