deps.ffmpeg: Fix libaom universal builds when using CMake 4
CMake 4.0 gained support for parallel installations, which means that
all install scripts (which can themselves call other install scripts
in subdirectories of the build tree) are gathered up and placed in a
JSON file, which is then used by CMake to run them in parallel.

Universal builds for incompatible projects "fake" the installation
procedure by copying a single architecture's build tree, use lipo
to combine the separate architectures into a final "fat" binary, and
finally run CMake to install this "fake" build tree.

Thus when using CMake 4, the paths in the corresponding JSON file all
also need to point to the files in the universal build tree.