Add qtserialport to obs-deps
The obs-ptz plugin uses qtserialport to communicate with remote PTZ
cameras. Unfortunatly, it does not seem possible to build QSerialPort
outside of the Qt6 library build environment because it links into
private Qt interfaces. The plugin used to maintain a private fork of
obs-deps just to add qtserialport, but that approach proved to be
fragile. Some users have had success putting a copy of qtserialport from
the official Qt binaries into the plugins directory, but again it is
easy to get out of sync.

Adding qtserialport to the list of Qt modules built in obs-deps solves
the problem with low complexity. The library is an official Qt
component, is small, and doesn't pull in any additional dependencies.