utils.pwsh: Remove manual pacman keyring upgrade
The Windows 11 Arm runners do not have MSYS2 installed, and attempts to
manually install it have failed. Installing via winget fails because
winget is not available on the windows-2022 or windows-11-arm runners.
Installing via chocolately fails because its install destination is
non-standard, and their option to change the install destination is a
paid option.

Just remove the manual pacman keyring upgrade for now.